refactor: change name impl to factory
