+++ src/main/factory/controllers/account/confirm-account-controller-factory.js
