+++ src/main/factory/controllers/task/delete-task-controller-factory.js
