+++ src/main/factory/middlewares/auth-middleware-factory.js
@@ -1,0 +1,13 @@
+const AuthMiddleware = require("../../../presentation/middlewares/auth-middleware")
+
+const { makeTokenGenerator } = require('../utils')
+
+const makeAuthenticateMiddleware = () => {  
+  return new AuthMiddleware({
+    tokenGenerator: makeTokenGenerator()
+  })
+}
+
+module.exports = {
+  makeAuthenticateMiddleware
+}
