+++ src/main/middlewares/auth.js
@@ -1,5 +1,5 @@
 const { adaptMiddleware } = require('../adapters/express-middleware-adapter')
-const { makeAuthenticateMiddleware } = require('../factory/middlewares/auth-middleware-impl')
+const { makeAuthenticateMiddleware } = require('../factory/middlewares/auth-middleware-factory')
 
 const auth = adaptMiddleware(makeAuthenticateMiddleware())
