+++ src/main/routes/account-router.js
@@ -4,11 +4,11 @@ const accountRouter = express.Router()
 const { adapt } = require('../adapters/express-router-adapter')
 const auth = require('../middlewares/auth')
 
-const { makeMeAccountController } = require('../factory/controllers/account/me-account-controller-impl')
-const { makeAddAccountController } = require('../factory/controllers/account/add-account-controller-impl')
-const { makeAuthenticateController } = require('../factory/controllers/account/authenticate-controller-impl')
-const { makeConfirmAccountController } = require('../factory/controllers/account/confirm-account-controller-impl')
-const { makeSendOtherMailConfirmAccountController } = require('../factory/controllers/account/send-otther-mail-confirm-account-controller-impl')
+const { makeMeAccountController } = require('../factory/controllers/account/me-account-controller-factory')
+const { makeAddAccountController } = require('../factory/controllers/account/add-account-controller-factory')
+const { makeAuthenticateController } = require('../factory/controllers/account/authenticate-controller-factory')
+const { makeConfirmAccountController } = require('../factory/controllers/account/confirm-account-controller-factory')
+const { makeSendOtherMailConfirmAccountController } = require('../factory/controllers/account/send-other-mail-confirm-account-controller-factory')
 
 accountRouter.get('/me', auth, adapt(makeMeAccountController()))
 accountRouter.post('/create', adapt(makeAddAccountController()))
