+++ src/main/routes/task-router.js
@@ -3,13 +3,13 @@ const taskRouter = express.Router()
 
 const { adapt } = require('../adapters/express-router-adapter')
 
-const { makeAddTaskController } = require('../factory/controllers/task/add-task-controller-impl')
-const { makeLoadTaskByIdController } = require('../factory/controllers/task/load-task-by-id-controller-impl')
-const { makeLoadTasksByUserIdController } = require('../factory/controllers/task/load-tasks-by-user-id-controller-impl')
-const { makeUpdateTaskByIdController } = require('../factory/controllers/task/update-task-by-id-controller-impl')
-const { makeDeleteTaskController } = require('../factory/controllers/task/delete-task-controller-impl')
-
 const auth = require('../middlewares/auth')
+
+const { makeAddTaskController } = require('../factory/controllers/task/add-task-controller-factory')
+const { makeLoadTaskByIdController } = require('../factory/controllers/task/load-task-by-id-controller-factory')
+const { makeLoadTasksByUserIdController } = require('../factory/controllers/task/load-tasks-by-user-id-controller-factory')
+const { makeUpdateTaskByIdController } = require('../factory/controllers/task/update-task-by-id-controller-factory')
+const { makeDeleteTaskController } = require('../factory/controllers/task/delete-task-controller-factory')
 
 taskRouter.get('/:id', auth, adapt(makeLoadTaskByIdController()))
 taskRouter.get('/all/:userId', auth, adapt(makeLoadTasksByUserIdController()))
